@@ -1,71 +1,44 @@
 {
   "name": "@colony/colony-starter-react",
-  "version": "1.0.0-beta.7",
+  "version": "1.0.0-beta.8",
   "description": "A simple starter project for building with Colony",
   "license": "MIT",
-  "engines": {
-    "node": "10.12.0"
-  },
   "publishConfig": {
     "access": "public"
   },
-  "scripts": {
-    "postinstall": "",
-    "start-ganache": "sh scripts/start_ganache.sh",
-    "deploy-contracts": "sh scripts/deploy_contracts.sh",
-    "start-trufflepig": "sh scripts/start_trufflepig.sh",
-    "seed-network": "node scripts/seed_network.js",
-    "build": "webpack -p --config webpack/production.config.js",
-    "start": "webpack-dev-server --config webpack/development.config.js",
-    "test": "jest --updateSnapshot"
+  "engines": {
+    "node": ">=10.12.0"
   },
-  "jest": {
-    "moduleNameMapper": {
-      "\\.scss$": "identity-obj-proxy"
-    },
-    "testURL": "http://localhost"
+  "files": ["/", ".gitignore"],
+  "scripts": {
+    "start-ganache": "colony service start-ganache",
+    "deploy-contracts": "colony service deploy-contracts",
+    "start-trufflepig": "colony service start-trufflepig",
+    "seed-network": "colony service seed-network",
+    "colony-setup": "node scripts/colonySetup.js",
+    "build": "react-scripts build",
+    "start": "react-scripts start",
+    "test": "react-scripts test --coverage --updateSnapshot",
+    "eject": "react-scripts eject"
   },
   "dependencies": {
+    "@colony/colony-cli": "^1.0.0-beta.0",
     "@colony/colony-js-client": "^1.11.2",
     "@colony/colony-js-contract-loader-http": "^1.10.0",
-    "@colony/purser-metamask": "^2.1.1",
-    "bn.js": "^4.11.8",
-    "buffer": "^5.1.0",
-    "ethers": "3.0.27",
-    "ipfs": "^0.29.2",
-    "react": "^16.6.1",
-    "react-dom": "^16.6.1",
-    "react-redux": "^5.1.1",
-    "react-router": "^4.3.1",
-    "react-router-dom": "^4.3.1",
-    "redux": "^4.0.1",
-    "redux-logger": "^3.0.6",
-    "web3": "^1.0.0-beta.35",
-    "web3-utils": "^1.0.0-beta.44"
+    "@colony/purser-metamask": "^2.2.0",
+    "@colony/purser-software": "^1.2.4",
+    "react": "^16.8.6",
+    "react-dom": "^16.8.6"
   },
   "devDependencies": {
-    "autoprefixer": "^8.6.2",
-    "babel-core": "^6.26.0",
-    "babel-jest": "^23.4.2",
-    "babel-loader": "^7.1.3",
-    "babel-polyfill": "^6.26.0",
-    "babel-preset-env": "^1.6.1",
-    "babel-preset-react": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "css-loader": "^0.28.11",
-    "extract-text-webpack-plugin": "^4.0.0-beta.0",
-    "identity-obj-proxy": "^3.0.0",
-    "jest": "^23.1.0",
     "node-sass": "^4.9.0",
-    "postcss-loader": "^2.1.5",
-    "precss": "^1.4.0",
-    "react-test-renderer": "^16.6.1",
-    "sass-loader": "^7.0.3",
-    "style-loader": "^0.21.0",
-    "trufflepig": "^1.1.1",
-    "webpack": "^4.12.1",
-    "webpack-cli": "^3.0.8",
-    "webpack-dev-server": "^3.1.5",
-    "webpack-merge": "^4.1.3"
-  }
+    "react-scripts": "^2.1.8",
+    "react-test-renderer": "^16.8.6"
+  },
+  "browserslist": [
+    ">0.2%",
+    "not dead",
+    "not ie <= 11",
+    "not op_mini all"
+  ]
 }
